Update / updated readme.mb and readme files
* Update / updated readme.mb and readme files

Why is this change neccesary?

To keep the readme file with the leatest info to install conekta ios sdk

How does it address the issue?

Updating the info in readme file

What side effects does this change have?

The merchants can install the latest version of conekta ios sdk following the readme